+++ test/functional/dd-functional-settlement-spec.js
@@ -141,6 +141,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     page.confirmVerification();
     page.setFederalPellGrants( 5500 );
     browser.sleep( 750 );
+    expect( page.pellGrantCapWarning.isDisplayed() ).toBeFalsy;
     expect( page.totalGrantsScholarships.getText() ).toEqual( '15,600' );
     expect( page.studentTotalCost.getText() ).toEqual( '28,026' );
     expect( page.remainingCostFinal.getText() ).toEqual( '-1,474' );
@@ -150,8 +151,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     page.confirmVerification();
     page.setFederalPellGrants( 10000 );
     browser.sleep( 750 );
-    // TODO: expect student is informed about the Pell Grant cap
-    // expect( EC.visibilityOf( page.pellGrantCapWarning ) );
+    expect( page.pellGrantCapWarning.isDisplayed() ).toBeTruthy;
     expect( page.totalGrantsScholarships.getText() ).toEqual( '15,915' );
     expect( page.studentTotalCost.getText() ).toEqual( '27,711' );
     expect( page.remainingCostFinal.getText() ).toEqual( '-1,789' );
@@ -270,6 +270,15 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalContributions.getText() ).toEqual( '19,000' );
     expect( page.remainingCostFinal.getText() ).toEqual( '1,526' );
+  } );
+
+  it( 'should properly update when both the Parent PLUS loan and the cash a student\'s family will provide are modified', function() {
+    page.confirmVerification();
+    page.setFamilyContribution( 4000 );
+    page.setParentPlusContribution( 2000 );
+    browser.sleep( 750 );
+    expect( page.totalContributions.getText() ).toEqual( '9,000' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '11,526' );
   } );
 
   it( 'should properly update when the work study earnings are modified within the allowed limit', function() {