+++ CHANGELOG.md
@@ -3,6 +3,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
 - Added string checking for program codes
+- Refine input of Pell Grants and family contributions
 
 ## 2.1.3
 - Added load tests